Remove unnecessary noise from CI breeze's output (#49952)
CI tests with breeze are running with `--verbose` options in order to
get more diagnostics when things fail. However there were a few
commmands always executed at the end of every command in breeze which
did not add any value for diagnostics and only generated noise:

* id retrievals for users and groups
* fixing ownership
* checking for rootless docker

Those commmands are now silecenced with `quiet=True` of run_command,
and the command nicely handles the case where we also capture output
of such command.